gateway/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
aiohttp>=3.8.0
pydantic>=1.8.0
PyJWT>=2.0.0
passlib>=1.7.4
python-multipart>=0.0.5
pydantic-settings>=2.0.0
PyYAML>=5.4.0
SQLAlchemy>=2.0.0
PyMySQL>=1.0.0
aiomysql>=0.2.0
greenlet>=1.0.0
email-validator>=1.1.3
stripe>=7.0.0
alembic>=1.7.0